remove control panel
